+++ src/main/java/com/leopoldo/bcv/services/interfaces/IRateScrapingService.java
@@ -1,6 +1,7 @@
 package com.leopoldo.bcv.services.interfaces;
+import java.math.BigDecimal;
 import java.util.Map;
 
 public interface IRateScrapingService {
-    Map<String, Double> scrapeRates() ;
+    Map<String, BigDecimal> scrapeRates() ;
 }
